backend/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==11.0.0
psycopg==3.2.3
psycopg2-binary==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3
typing_extensions==4.12.2
django-filter==24.3
django-cors-headers==4.6.0

